Внутренние вычисления переведены в формат decimal для большей точности. Также добавлено округление вверх до тысячных при рассчёте потоков, чтобы нивелировать неточности, возникающие при работе с нечётными дробями.

+++ Model/Elements/ResourceOwerflow.cs
@@ -5,10 +5,10 @@ namespace SatisfactoryProductionManager.Model.Elements
     public class ResourceOverflow
     {
         public string Resource { get; set; }
-        public double CountPerMinute { get; set; }
+        public decimal CountPerMinute { get; set; }
 
 
-        public ResourceOverflow(string resource, double count)
+        public ResourceOverflow(string resource, decimal count)
         {
             Resource = resource;
             CountPerMinute = count;
